Divided the functions of the CardPricing and the slider
